server/requirements.txt
# Core Flask packages
Flask
Flask-CORS
Flask-Bcrypt
Flask-JWT-Extended
Flask[async]
flask_APScheduler
Flask-SQLAlchemy
Flask-Migrate

# Database
pymongo
psycopg2
chromadb

# LangChain ecosystem (updated versions)
langchain
langchain-core
langchain-community
langchain-experimental
langchain-openai
langchain-pinecone
langchain-ollama
langchain_text_splitters

# Document processing
unstructured[all-docs]
pdfplumber

# Embeddings and ML
fastembed
sentence-transformers
numpy
pinecone-client

# API and HTTP
openai
Requests
urllib3
ollama
elevenlabs
psycopg2

# Utils
python-dotenv
pydantic
pydantic-settings
google-auth
hypercorn
tenacity
marshmallow

sendgrid
pytest
pytest-flask
pytest-cov